Added approximate carving of encrypted rar files
Because rar files do not contain a total length field, and when headers are
encrypted it is impossible even to iterate over good blocks to find the end of
a file, a simple heuristic is used to guess where encrypted data ends and that
length is passed to the carver.

Currently that heuristic is simply to stop when four identical bytes are
encountered.