reorganize python library module for re-use
Previously opening a BulkExtractor object a second time would crash by
attempting to modify scanner configuration a second time.  Now a
module-level initialization system is in place and many separate Session
objects may be used in sequence.

Organization of the module source is also improved; low-level constants
are moved to the bottom with other low-level code.  Pydoc is updated to
be more comprehensive.