GH Workflows: Create CI job for Coverity scan
Coverity Scan is a static source code analyzer to find bugs, security vulnerabilities, and other potential issues. This PR adds a GitHub Workflow CI job to create a new Coverity Scan build daily, to ease detection and remediation of software bugs.

Before this is merged in, the following steps are required:
1) Register the "bulk_extractor" project on Coverity's website here: https://scan.coverity.com/projects/ (it would be best if an official maintainer did this)
2) Create two new secrets:
a) `COVERITY_SCAN_EMAIL` with the email address that should be notified with the scan results. It is not possible to disable the scan result emails.
b) `COVERITY_SCAN_TOKEN` with the Project token from the Coverity project page (e.g., https://scan.coverity.com/projects/bulk_extractor?tab=project_settings )

Optionally, you may also wish to:
1) Configure Coverity to ignore directories that you don't want it to scan. (e.g., the `/src/be20_api/.*` directory)

After this PR is merged, GitHub will submit builds to Coverity with the schedule specified in the yaml file. However, if desired, you can trigger a manual run by going to `Actions -> Coverity Scan`, and clicking on the `Run workflow` button.